requirements.txt
flask
joblib
numpy
pandas
python-dotenv
scikit-learn
google-generativeai
gTTS